refactor: split make queue
