+++ crates/rspack_core/src/compiler/mod.rs
@@ -20,7 +20,7 @@ use tracing::instrument;
 
 pub use self::compilation::*;
 pub use self::hmr::{collect_changed_modules, CompilationRecords};
-pub use self::make::{FactorizeTask, MakeParam};
+pub use self::make::MakeParam;
 pub use self::module_executor::{ExecuteModuleId, ModuleExecutor};
 use crate::cache::Cache;
 use crate::tree_shaking::symbol::{IndirectType, StarSymbolKind, DEFAULT_JS_WORD};
